git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@1884 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CCameraSceneNode.cpp
@@ -230,7 +230,7 @@ void CCameraSceneNode::OnRegisterSceneNode()
 
 	f32 dp = tgtv.dotProduct(up);
 
-	if ( core::equals ( fabs ( dp ), 1.f ) )
+	if ( core::equals(fabsf(dp), 1.f) )
 	{
 		up.X += 0.5f;
 	}
